moved function. token to localstorage
login and register is moved to own file.
auth token and name is added to local storage if login is successfull.
